Explore: move new function to helpers
